# apps/docs/static/openapi/framework-live.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/cms/get-entry": {
      "get": {
        "operationId": "CmsController_getEntry",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "CMS entry identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns CMS entry."
          }
        },
        "summary": "Get CMS entry",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/get-entries": {
      "get": {
        "operationId": "CmsController_getEntries",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "description": "CMS content type key (for example `page`, `article`).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filters",
            "required": false,
            "in": "query",
            "description": "Optional map of integration-specific filters.",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns CMS entries."
          }
        },
        "summary": "Get CMS entries",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/page": {
      "get": {
        "operationId": "CmsController_getPage",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "slug",
            "required": true,
            "in": "query",
            "description": "Page slug identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns CMS page."
          }
        },
        "summary": "Get CMS page",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/pages": {
      "get": {
        "operationId": "CmsController_getPages",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns CMS pages."
          }
        },
        "summary": "Get CMS pages",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/login-page": {
      "get": {
        "operationId": "CmsController_getLoginPage",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns CMS login page."
          }
        },
        "summary": "Get CMS login page",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/not-found-page": {
      "get": {
        "operationId": "CmsController_getNotFoundPage",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns CMS not-found page."
          }
        },
        "summary": "Get CMS not-found page",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/header": {
      "get": {
        "operationId": "CmsController_getHeader",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "Header content identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns CMS header."
          }
        },
        "summary": "Get CMS header",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/footer": {
      "get": {
        "operationId": "CmsController_getFooter",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "Footer content identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns CMS footer."
          }
        },
        "summary": "Get CMS footer",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/app-config": {
      "get": {
        "operationId": "CmsController_getAppConfig",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "referrer",
            "required": false,
            "in": "query",
            "description": "Optional referrer URL or route used for context-aware app config.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns CMS app config."
          }
        },
        "summary": "Get CMS app config",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/blocks/faq": {
      "get": {
        "operationId": "CmsController_getFaqBlock",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "CMS entry identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns FAQ block."
          }
        },
        "summary": "Get FAQ block",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/blocks/ticket-list": {
      "get": {
        "operationId": "CmsController_getTicketListBlock",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "CMS entry identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns ticket-list block."
          }
        },
        "summary": "Get ticket-list block",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/blocks/ticket-details": {
      "get": {
        "operationId": "CmsController_getTicketDetailsBlock",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "CMS entry identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns ticket-details block."
          }
        },
        "summary": "Get ticket-details block",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/blocks/notification-list": {
      "get": {
        "operationId": "CmsController_getNotificationListBlock",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "CMS entry identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns notification-list block."
          }
        },
        "summary": "Get notification-list block",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/blocks/notification-details": {
      "get": {
        "operationId": "CmsController_getNotificationDetailsBlock",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "CMS entry identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns notification-details block."
          }
        },
        "summary": "Get notification-details block",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/blocks/article-list": {
      "get": {
        "operationId": "CmsController_getArticleListBlock",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "CMS entry identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns article-list block."
          }
        },
        "summary": "Get article-list block",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/blocks/article-details": {
      "get": {
        "operationId": "CmsController_getArticleDetailsBlock",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "CMS entry identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns article-details block."
          }
        },
        "summary": "Get article-details block",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/blocks/invoice-list": {
      "get": {
        "operationId": "CmsController_getInvoiceListBlock",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "CMS entry identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns invoice-list block."
          }
        },
        "summary": "Get invoice-list block",
        "tags": [
          "cms"
        ]
      }
    },
    "/api/cms/blocks/invoice-details": {
      "get": {
        "operationId": "CmsController_getInvoiceDetailsBlock",
        "parameters": [
          {
            "name": "preview",
            "required": false,
            "in": "query",
            "description": "When true, returns draft/preview CMS content.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "CMS entry identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "query",
            "description": "Locale code used to localize the response, for example `pl` or `en`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns invoice-details block."
          }
        },
        "summary": "Get invoice-details block",
        "tags": [
          "cms"
        ]
      }
    },
...
  },
  "info": {
    "title": "O2S API Harmonization HTTP API",
    "description": "Live API documentation generated from current NestJS controllers and DTOs.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}